lms-rangga-master/requirements.txt
asgiref==3.5.0
Django==4.0.1
mysql-connector==2.2.9
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
